Xcm config to a separate file: moonriver
